Move the channels into Socket
The library should be agnostic to channels and multiple threads, meaning
that we give the user the ability to use it single-threadedly, or to
spawn threads which define their own queues that other threads can use.

This patch is the first step in that direction.

We do lose a blocking `recv` right now, but I think we can do that in a
smarter way, instead of `manual_poll` doing recv, we can let `recv` do a
direct receive on the socket instead.